Created radar_utils.py in pyart.util. Moved join_radar function to radar_utils.
